# Perf improvement - Inline DSL attributes