@DocX 2017-06-23
There is definitely more room to go - inline other serialized fields like `meta`, `id`, ... - And perhaps then inline the whole `data` map for given object. Not sure how far this can be stretched before it is "needed" to go runtime.

Let me know what you think.