@alanpeabody 2017-06-23
@DocX thanks a bunch! I have tried doing something like this a couple of times and fallen short, so much so that I had been just suggesting not using the attributes macro unless you really need it.

I am going to take a bit of time and study this, but I will likely merge this weekend.